Added a few more comments
Added a note about equality in the README, since I have noted Date can be guaranteed to be equal with built-in equality, avoiding the need for using the `.Equal` method.

I considered adding the note for the `.Equal` method, but since it actually checks in a different manner, I left it as is, but effectively `==` and `.Equal` should always return the same (though some fuzzy tests on this for the future would be appropriate).

Also noted that the `.Add` method is not recommended to use, as its behaviour is a bit opaque due to dealing with dates, and that `AddDate` should always be preferred instead.

To ensure the guarantee that dates are always comparable, I noted that's the reason why the type is a struct with a `time.Time` rather than just straight a `time.Time`, so a `time.Time` cannot be cast to be a `date.Date`, and `date.NewDateFromTime` must be used.